HPCC-34155: Move index event hotspot to shared library
