+++ common/eventconsumption/eventindexhotspot.h
@@ -17,9 +17,23 @@
 
 #pragma once
 
-#include "evtool.hpp"
-#include "jevent.hpp"
-#include "jstream.hpp"
+#include "eventconsumption.h"
+#include "eventoperation.h"
+#include "eventvisitor.h"
 
-extern IEvToolCommand* createIndexSummaryCommand();
-extern IEvToolCommand* createIndexHotspotCommand();
+constexpr static byte defaultGranularityBits = 0; // 1 page per bucket
+
+class CIndexHotspotOp : public CEventConsumingOp
+{
+public:
+    virtual bool ready() const override;
+    virtual bool doOp() override;
+public:
+    void setObservedEvent(EventType _observedEvent);
+    void setGranularity(byte bits);
+    void setLimit(byte _limit);
+protected:
+    EventType observedEvent{EventNone};
+    byte granularityBits{defaultGranularityBits};
+    byte limit{10};
+};
